+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - `fusion` method for tilings based on the `Fusion` class.
+- `row_and_column_separation` method for tilings based on the
+  `RowColSeperation` class.
+
+### Changed
+- Update permuta to 1.2.1
+- Update comb_spec_searcher to 0.2.2
 
 ## [1.0.1] - 2019-08-26
 ### Changed